apps/api/apps/deals/api/urls.py
from django.urls import include
from rest_framework.routers import DefaultRouter
from .views import DealViewSet

router = DefaultRouter()
router.register('deals', DealViewSet, basename='deal')
urlpatterns = router.urls
